feat: changed strategy to console app
