+++ teardrop/takecare.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.6.33829.357
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "takecare", "takecare\takecare.csproj", "{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "teardrop", "teardrop\teardrop.csproj", "{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
