--- teardrop/takecare/Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Windows.Forms;
-
-namespace takecare
-{
-    static class Program
-    {
-        [STAThread]
-        static void Main()
-        {
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new FrmMain());
-        }
-    }
-}
