--- teardrop/takecare/Properties/Settings.settings
@@ -1,51 +1,0 @@
-﻿<?xml version='1.0' encoding='utf-8'?>
-<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="teardrop.Properties" GeneratedClassName="Settings">
-  <Profiles />
-  <Settings>
-    <Setting Name="key" Type="System.String" Scope="User">
-      <Value Profile="(Default)" />
-    </Setting>
-    <Setting Name="extension" Type="System.String" Scope="User">
-      <Value Profile="(Default)">.toasty</Value>
-    </Setting>
-    <Setting Name="db_host" Type="System.String" Scope="User">
-      <Value Profile="(Default)">127.0.0.1</Value>
-    </Setting>
-    <Setting Name="db_user" Type="System.String" Scope="User">
-      <Value Profile="(Default)">root</Value>
-    </Setting>
-    <Setting Name="db_pass" Type="System.String" Scope="User">
-      <Value Profile="(Default)" />
-    </Setting>
-    <Setting Name="db_database" Type="System.String" Scope="User">
-      <Value Profile="(Default)">teardrop</Value>
-    </Setting>
-    <Setting Name="db_enable" Type="System.Boolean" Scope="User">
-      <Value Profile="(Default)">False</Value>
-    </Setting>
-    <Setting Name="debug" Type="System.Boolean" Scope="User">
-      <Value Profile="(Default)">False</Value>
-    </Setting>
-    <Setting Name="disable_taskmgr" Type="System.Boolean" Scope="User">
-      <Value Profile="(Default)">False</Value>
-    </Setting>
-    <Setting Name="clickMouse" Type="System.Boolean" Scope="User">
-      <Value Profile="(Default)">False</Value>
-    </Setting>
-    <Setting Name="message" Type="System.String" Scope="User">
-      <Value Profile="(Default)">&lt;h1&gt;Github:&lt;/h1&gt;&lt;p&gt;hackthedev/teardrop&lt;/p&gt;</Value>
-    </Setting>
-    <Setting Name="unkillable" Type="System.Boolean" Scope="User">
-      <Value Profile="(Default)">False</Value>
-    </Setting>
-    <Setting Name="application_name" Type="System.String" Scope="User">
-      <Value Profile="(Default)" />
-    </Setting>
-    <Setting Name="theme" Type="System.String" Scope="User">
-      <Value Profile="(Default)">flash</Value>
-    </Setting>
-    <Setting Name="application_title" Type="System.String" Scope="User">
-      <Value Profile="(Default)" />
-    </Setting>
-  </Settings>
-</SettingsFile>
