--- teardrop/takecare/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="DeviceId" version="6.6.0" targetFramework="net48" />
-</packages>
